Refactor setup.py jsdeps
